Give up on bulk insert!
Updated the database using create! methods instead of the new
insert_all! method. I think the way to make this work is to add a
unique token column to the questions table and use that in the
unique_by parameter of insert_all. It's too much hassle for now. I
will try it out later.