Fix an MUI error: changing uncontrolled value state of RadioGroup to controlled
From error message: Elements shouldn't uncontrolled to controlled
or vice versa. Nature of state determined during first render.
controlled if state value != undefined. This was fixed by not
setting the value prop for the radio group and using only onChange
to track form state instead. Supposedly refs are better to work
with when dealing with forms.