Persistent user sessions in the browser
This solves the user login session being broken on page refresh.
It uses react-cookie package to save the session token in a
cookie that expires 30 mins after login. The session implementation
needs to be hardened. As of now the session tokens are stored
directly in a db table, leaving them vulnerable. Instead a session
digest needs to be saved in db using one of rails's encryption
packages. Also logic needs to be implemented to refresh server
tokens.